verifying login page works
