+++ src/pages/LoginPage/index.js
@@ -1,24 +1,12 @@
 import React, { useState } from "react";
+import LoginForm from "../../components/LoginForm";
 // import { Username_Password, Email, Submit  } from "../../components";
 //forms for login page
-const LoginPage = () => {
-    //allows form to set username and password
-    const [Name, setName] = useState("");
-    const [Password, setPassword] = useState("");
-
-    const handleSubmit = async () => {
-        const user = {
-            Name,
-            Password,
-        };
-        console.log(user);
-        //fill with relevant code when hooked to back-end user api.
-    };
-    
+const LoginPage = () => {    
     //what actually gets rendered on the page
     return (
-        <div className="login-form">
-            <div>Testing LoginPage</div>
+        <div className="login-page">
+            <LoginForm />
         </div>
     )
 }
